Finalizadas barras de vida

+++ client/engine/components/visual_character_component.cpp
@@ -237,8 +237,8 @@ void VisualCharacterComponent::update() {
         }
     }
     StatsComponent& stats = entity->get_component<StatsComponent>();
-    health_bar.set_fill_proportion(stats.get_stat_current_value("hp") /
-                                   stats.get_stat_max_value("hp"));
+    health_bar.set_fill_proportion((float)stats.get_stat_current_value("hp") /
+                                   (float)stats.get_stat_max_value("hp"));
 }
 
 bool VisualCharacterComponent::is_moving() {
